requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0